amblyopia_backend/requirements.txt
# =============================================================================
# Amblyopia Care System — Python Requirements (LOCAL, NO DOCKER)
# Aravind Eye Hospital, Coimbatore, India
# =============================================================================

# ── FastAPI / Web ─────────────────────────────────────────────────────────────
fastapi==0.115.8
uvicorn[standard]==0.34.0
starlette==0.45.3
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.3

# ── Auth / Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
bcrypt==4.1.2

# ── Config / Utils ────────────────────────────────────────────────────────────
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytz==2023.3.post1

# ── Redis (session cache, offline sync queue) ─────────────────────────────────
redis==5.0.1

# ── ML / Vision ───────────────────────────────────────────────────────────────
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ── Audio processing ─────────────────────────────────────────────────────────
librosa>=0.10.0
soundfile>=0.12.1
openai-whisper>=20231117
webrtcvad==2.0.10          # optional — graceful skip if fails

# ── MLflow (LOCAL file store — no server required) ────────────────────────────
mlflow>=2.8.0

# ── Reporting / QR ───────────────────────────────────────────────────────────
reportlab>=4.0.0
qrcode>=7.4.0

# ── Observability (optional) ──────────────────────────────────────────────────
prometheus-fastapi-instrumentator>=6.1.0
sentry-sdk>=1.38.0         # set SENTRY_DSN= to disable

# ── Notifications (optional — leave keys blank to disable) ───────────────────
twilio>=8.0.0
firebase-admin>=6.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0

# ── Integration-layer extras (installed by setup/install_all.sh) ──────────────
# torch torchvision torchaudio  → bash setup/install_all.sh
# realesrgan basicsr facexlib gfpgan
# ultralytics
# minio                         → optional, set MINIO_DISABLED=true to skip
